Extract shared stats method
