+++ src/adapter/10/stats.ts
@@ -18,23 +18,6 @@ export function getDiffType(child: VNode, prev: DiffType) {
 	return prev;
 }
 
-export function updateDiffStats(
-	stats: Stats,
-	diff: DiffType,
-	childCount: number,
-) {
-	if (diff === DiffType.KEYED) {
-		stats.keyed.total++;
-		stats.keyed.children.push(childCount);
-	} else if (diff === DiffType.UNKEYED) {
-		stats.unkeyed.total++;
-		stats.unkeyed.children.push(childCount);
-	} else if (diff === DiffType.MIXED) {
-		stats.mixed.total++;
-		stats.mixed.children.push(childCount);
-	}
-}
-
 export function recordComponentStats(
 	config: RendererConfig10,
 	stats: Stats,
